misc/requirements.txt
# Life Strands System - Python Dependencies

# ================================
# Core Framework Dependencies
# ================================

# Web framework and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database and ORM
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0

# Redis for caching and queuing
redis[hiredis]==5.0.1
aioredis==2.0.1

# ================================
# Machine Learning and NLP
# ================================

# LLM inference
llama-cpp-python==0.2.11
torch>=2.0.0,<3.0.0
transformers==4.35.0
tokenizers==0.14.1

# Vector embeddings and similarity search
sentence-transformers==2.2.2
numpy>=1.24.0,<2.0.0
faiss-cpu==1.7.4

# Text processing
nltk==3.8.1
spacy>=3.7.0,<4.0.0

# ================================
# Database and Vector Extensions
# ================================

# PostgreSQL with vector support
psycopg2-binary==2.9.9
pgvector==0.2.3

# JSON schema validation
jsonschema==4.19.2

# ================================
# Web and Networking
# ================================

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# WebSocket support
websockets==12.0
python-socketio==5.10.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ================================
# Monitoring and Logging
# ================================

# Structured logging
structlog==23.2.0
python-json-logger==2.0.7

# Metrics and monitoring
prometheus-client==0.19.0
psutil==5.9.6

# GPU monitoring
GPUtil==1.4.0
nvidia-ml-py==12.535.133

# Tracing (optional)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger==1.21.0

# ================================
# Development and Testing
# ================================

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-autoapi==3.0.0

# ================================
# Utility Libraries
# ================================

# Configuration management
python-dotenv==1.0.0
click==8.1.7

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# Data validation and serialization
marshmallow==3.20.1
cerberus==1.3.5

# File handling
aiofiles==23.2.1
pathlib2==2.3.7

# Cryptography
cryptography==41.0.7

# ================================
# Performance and Optimization
# ================================

# Async utilities
asyncio-throttle==1.0.2
aiocache==0.12.2

# Memory profiling
memory-profiler==0.61.0
pympler==0.9

# Performance monitoring
py-spy==0.3.14

# ================================
# External Service Integrations
# ================================

# Cloud storage (optional)
boto3==1.34.0
azure-storage-blob==12.19.0

# Email support (optional)
aiosmtplib==3.0.1

# Webhooks
requests==2.31.0

# ================================
# Model-Specific Dependencies
# ================================

# GGUF model support
gguf==0.4.0

# Additional ML utilities
scikit-learn==1.3.2
scipy>=1.10.0,<2.0.0

# Image processing (if needed for multimodal)
Pillow>=9.0.0,<11.0.0

# ================================
# Production Dependencies
# ================================

# Process management
gunicorn==21.2.0
supervisor==4.2.5

# Environment management
python-decouple==3.8

# Health checks
httptools==0.6.1

# ================================
# Optional Development Tools
# ================================

# API documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Profiling
line-profiler==4.1.1
cProfile-to-flamegraph==0.1.0

# Database migrations
yoyo-migrations==8.2.0

# Code formatting
autopep8==2.0.4
docformatter==1.7.5

# ================================
# Version Constraints and Compatibility
# ================================

# Ensure compatibility
setuptools>=65.0.0
wheel>=0.37.0
pip>=21.0.0

# Python version compatibility markers
typing-extensions>=4.0.0; python_version<"3.10"
importlib-metadata>=4.0.0; python_version<"3.10"

# ================================
# Security and Compliance
# ================================

# Security scanning
safety==2.3.5
bandit==1.7.5

# Dependency management
pip-audit==2.6.1

# ================================
# Platform-Specific Dependencies
# ================================

# Windows-specific
pywin32>=227; sys_platform=="win32"
wmi>=1.5.0; sys_platform=="win32"

# Linux-specific performance tools
setproctitle==1.3.3; sys_platform=="linux"

# ================================
# Development Environment
# ================================

# Jupyter support (optional)
jupyter==1.0.0
ipykernel==6.26.0

# IDE support
python-lsp-server==1.9.0
rope==1.11.0

# ================================
# Build and Distribution
# ================================

# Package building
build==1.0.3
twine==4.0.2

# Container support
docker==6.1.3

# ================================
# Pinned Versions for Stability
# ================================

# Core async libraries - pinned for stability
anyio==4.0.0
sniffio==1.3.0

# Network libraries
urllib3>=1.26.0,<3.0.0
certifi>=2022.12.7

# Serialization
orjson==3.9.10  # Fast JSON serialization
msgpack==1.0.7

# ================================
# Optional GPU Dependencies
# ================================

# CUDA support (uncomment if using GPU)
# torch-audio==2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.15.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# AMD ROCm support (uncomment if using AMD GPU)
# torch-audio==2.0.0+rocm5.4.2 --index-url https://download.pytorch.org/whl/rocm5.4.2
# torchvision==0.15.0+rocm5.4.2 --index-url https://download.pytorch.org/whl/rocm5.4.2

# ================================
# Environment-Specific Overrides
# ================================

# Development only
pytest-xdist==3.4.0  # Parallel test execution
pytest-benchmark==4.0.0  # Performance benchmarks
factory-boy==3.3.0  # Test data generation

# Production optimizations
uvloop==0.19.0; sys_platform!="win32"  # Fast event loop (Unix only)
orjson==3.9.10  # Fast JSON parsing